+++ CHANGES.md
@@ -11,7 +11,7 @@
 
 - Added support for I3DM 3D Tile content files.
 - Added `forEachNodeInScene` to `CesiumGltf::Model`.
-- Added `removeUnusedBuffers` to `GltfUtilities`.
+- Added `removeUnusedBuffers`, `removeUnusedMeshes`, and `removeUnusedMaterials` to `GltfUtilities`.
 - Added the following new methods to the `Uri` class: `unescape`, `unixPathToUriPath`, `windowsPathToUriPath`, `nativePathToUriPath`, `uriPathToUnixPath`, `uriPathToWindowsPath`, and `uriPathToNativePath`.
 - Added `LayerWriter` to the `CesiumQuantizedMeshTerrain` library and namespace.
 - Drastically improved the performance of `GltfUtilities::collapseToSingleBuffer` for glTFs with many buffers and bufferViews.